Remove admission webhook (#44)
As external-dns is able to handle changes to routing-weight annotations
correctly, eg. simple to weighted records, there is no need for an admission
webhook in this controller. Although it gives us some consistency in the objects
it also adds complexity to the deployment and requires higher uptime in general
as Ingress resources cannot be created if the webhook is down.

This change removes the admission webhook controller along with reverts the
config resources to not enable cert-manager and webhooks.